Session 37 — OQ-85: widen champion_credit to u128 (cap 8→16)
Real Stack M matches accumulate >8 distinct combo-tick targets in a single
turn (multi-Tempest turns, plus Tempest ticks any neighbour it pushes —
friend or foe). The 8-slot tracked_enemies array was sized assuming
enemies-only and packed cross-product into u64. Release builds stripped
the debug_assert! guarding the array write and proceeded to UB.

- MAX_TRACKED_ENEMIES 8 → 16; champion_credit u64 → u128. Casters stay at
  8 (covers 5 Champions + King = 6 max).
- ensure_tracked_enemy/_caster now assert! (panic in release) instead of
  debug_assert!, to prevent silent OOB if the cap is ever hit again.
- Gauntlet's 4 real-match tests un-gated; all 7 pass in release.
- tier2_against_self_does_not_pass asserted >=3 games but mirrored_bo3
  early-returns at 2-0; corrected to >=2.
- OQ-85 marked resolved in design.db.

+++ game/crates/core_engine/src/game_logic/action.rs
@@ -439,7 +439,7 @@ pub struct Undo {
     pub p2_money_delta: i16,
 
     /// Combo-credit + tracked-enemies + tracked-casters snapshot.
-    pub prev_champion_credit: u64,
+    pub prev_champion_credit: u128,
     pub prev_tracked_enemies: [u8; crate::state::position::MAX_TRACKED_ENEMIES],
     pub prev_tracked_enemies_len: u8,
     pub prev_tracked_casters: [u8; crate::state::position::MAX_TRACKED_CASTERS],
